+++ config/splat.us.stno3.yaml
@@ -129,8 +129,7 @@ segments:
       - [0x18838, .data, tile_data] # tile data
       - [0x26638, .data, tile_data] # tile definitions
       - [0x2EA68, .data, gen/sprites]
-      - [0x373E0, data]
-      - [0x373E8, .rodata, entrance_stage_entities] # EntityCastleDoor
+      - [0x373E0, .rodata, entrance_stage_entities] # EntityCastleDoor
       - [0x37400, .rodata, entrance_stage_entities] # EntityStairwayPiece
       - [0x37414, .rodata, 3C4EC]
       - [0x3742C, .rodata, 3E134] #EntityFlyingOwlAndLeaves
