+++ config/splat.us.stnp3.yaml
@@ -132,8 +132,7 @@ segments:
       - [0x168F4, .data, tile_data] # tile data
       - [0x21CF4, .data, tile_data] # tile definitions
       - [0x2A124, .data, gen/sprites]
-      - [0x31EA0, data]
-      - [0x31EA8, .rodata, entrance_stage_entities] # EntityStairwayPiece
+      - [0x31EA0, .rodata, entrance_stage_entities] # EntityStairwayPiece
       - [0x31EBC, .rodata, e_explosion_puff_opaque] # EntityExplosionPuffOpaque
       - [0x31ED0, .rodata, slogra] # EntitySlogra .rodata, 36990
       - [0x31EDC, .rodata, gaibon] # EntityGaibon .rodata, 36990
